requirements.txt

asgiref==3.8.1
Django

djangorestframework==3.15.2
jalali_core==1.0.0
jdatetime==5.0.0
setuptools==71.0.3
sqlparse==0.5.1
